Add amplitude normalization, logarithmic modulatable slider, and waveform param modulation
Fix EMA amplitude squish by rescaling smoothed buffer peak to match raw
waveform each frame. Add ModulatableDrawableSliderLog wrapper (flags
passthrough to ImGui::SliderFloat). Convert all waveform geometry params
(radius, height, thickness, smooth, motion) to modulatable sliders and
register in both drawable_params and param_registry field table.

+++ src/config/drawable_config.h
@@ -20,7 +20,7 @@ struct DrawableBase {
 
 struct WaveformData {
     float amplitudeScale = 0.35f;
-    int thickness = 2;
+    float thickness = 2.0f;
     float smoothness = 5.0f;
     float radius = 0.25f;
     float waveformMotionScale = 1.0f;
